Logging changes and synaptic transitions
Synaptic transitions occur with this set of parameters. 
Logging updated to plot two values of weight at each plastic update:
after drift is added, and after the update is finished. The after drift
value is plotted at previous time, but that's close enough. 
There also may be a bug: sometimes there seems to be some kind of update
without any data in the log file.