Starting Fig 2-c from Brader, Senn, Fusi (2007)
I need this because V behaves differently than in the original 
paper, and this will help find a set of parameters for which 
LTP and LTD transitions happen. I only consider pre-synaptic 
rate = 50Hz. 
There is an issue with poisson source (it always 
generates the same spike train for the same rate).