Fix a lot of minor errors in integration tests
The minor errors were turning the directory into an ocean of red in my
IDE, despite there being very few actual errors. (For the most part, the
issues are that Neo is an unanalysable pile of rubbish. A secondary
issue is potential global/local variable masking that doesn't actually
kick in but looks like it does to code analysis tools; I've not been
consistent in chasing these down since they were just warnings.)

Errors that remain are real and not obvious how to fix.